+++ cmd/agent/agent_test.go
@@ -3,6 +3,7 @@ package main
 import (
 	"testing"
 
+	"github.com/netzen86/collectmetrics/internal/agent"
 	"github.com/netzen86/collectmetrics/internal/api"
 )
 
@@ -19,7 +20,7 @@ func TestCollectMetrics(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			CollectMetrics(tt.args.metrics, tt.args.counter)
+			agent.CollectMetrics(tt.args.metrics, tt.args.counter)
 		})
 	}
 }
